some cleaning and renaming generators generate pockets directly selectors/ selection modifiers modify which generator gets selected or how it gets selected
